+++ src/protocol/functions/Preset_Multiple_Registers.js
@@ -21,7 +21,7 @@ var PresetMultipleRegister = function (pdu) {
         respPDU.modbus_function = pdu.modbus_function | 0x80;
         respPDU.modbus_data[0] = 0x02;
 
-        this.emit('modbus_exception','ILLEGAL DATA ADDRESS');
+        //this.emit('modbus_exception','ILLEGAL DATA ADDRESS');
     }
     else {
         let values = new Map();