+++ src/protocol/functions/Read_Coil_Status.js
@@ -23,7 +23,7 @@ var ReadCoilStatus = function (pdu) {
         respPDU.modbus_function = pdu.modbus_function | 0x80;
         respPDU.modbus_data[0] = 0x02;
 
-        this.emit('modbus_exception','ILLEGAL DATA ADDRESS');
+        //this.emit('modbus_exception','ILLEGAL DATA ADDRESS');
     }
     else {
         //cantidad de coils a leer